requirements.txt
Pillow
numpy
requests
python-dotenv
